python_sr_morse_text/pyproject.toml
[project]
name = "python-sr-morse-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faust-streaming>=0.10.14",
    "SpeechRecognition>=3.10.0",
    "numpy>=1.26.0",
    "pyaudio>=0.2.14",
    "protobuf>=6.33.2",
]
